+++ examples/00-static-files.lisp
@@ -10,11 +10,7 @@
     ;; NOTE: The `ev-data' comes in "naked", so we must do a pointer cast (this
     ;; is free) to the real type we require.
     (let ((hm   (cast ev-data (* http-message)))
-          (opts (make-alien http-serve-opts)))
-      (setf (slot opts 'root-dir) ".")
-      ;; NOTE: Notice the need to explicitly set this to NULL. This is a
-      ;; Mongoose setting that ensures Posix will be used as the filesystem.
-      (setf (slot opts 'fs) nil)
+          (opts (http-serve-opts ".")))
       (http-serve-dir c hm opts)
       (free-alien opts))))