+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Mongoose
 
+## Unreleased
+
+#### Added
+
+- `http-serve-opts` helper for initializing the fields of the struct of the same
+  name in a useful way.
+
 ## 0.1.0 (2026-01-21)
 
 Initial release, including just what's necessary to run an HTTP server. TLS will